2/2 add SUM vs %

+++ src/main/java/com/example/patrol_be/dto/DivisionSummaryDTO.java
@@ -9,31 +9,31 @@ import lombok.*;
 public class DivisionSummaryDTO {
     private String division;
 
-    private long allTtl;
-    private long allI;
-    private long allII;
-    private long allIII;
-    private long allIV;
-    private long allV;
+    private double allTtl;
+    private double allI;
+    private double allII;
+    private double allIII;
+    private double allIV;
+    private double allV;
 
-    private long proDoneTtl;
-    private long proDoneI;
-    private long proDoneII;
-    private long proDoneIII;
-    private long proDoneIV;
-    private long proDoneV;
+    private double proDoneTtl;
+    private double proDoneI;
+    private double proDoneII;
+    private double proDoneIII;
+    private double proDoneIV;
+    private double proDoneV;
 
-    private long hseDoneTtl;
-    private long hseDoneI;
-    private long hseDoneII;
-    private long hseDoneIII;
-    private long hseDoneIV;
-    private long hseDoneV;
+    private double hseDoneTtl;
+    private double hseDoneI;
+    private double hseDoneII;
+    private double hseDoneIII;
+    private double hseDoneIV;
+    private double hseDoneV;
 
-    private long remainTtl;
-    private long remainI;
-    private long remainII;
-    private long remainIII;
-    private long remainIV;
-    private long remainV;
+    private double remainTtl;
+    private double remainI;
+    private double remainII;
+    private double remainIII;
+    private double remainIV;
+    private double remainV;
 }
